// src/types/dashboard/Ecommerce.ts
/*Weekly stats*/
type weeklyStats = {
    title: string;
    subtitle: string;
    rank: string;
    bgcolor: string;
    textcolor: string;
};

/*Payment Gatway*/
type paymentGateway = {
    img: string;
    title: string;
    subtitle: string;
    rank: string;
    disable: string;
    bgcolor: string;
};

export { weeklyStats, paymentGateway }